Neaten file structure for samplers and wrappers.
