+++ chrono_samplers/chrono_sampler_exact.h
@@ -38,52 +38,51 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 ** File Details **
 
-Filename: sampler_pseudo_nn.h
-Created: Oct 10, 2014
+Filename: chrono_sampler_exact.h
+Created: Feb 19, 2014
 Author: Y. H. Tang
-Description: Spatial sampler that provides a value at a point
-             using a pseudo nearest neighbour interpolation.
+Description: Temporal sampler that samples at exactly the time specified and performs no
+             interpolation
 */
 
-#ifndef MUI_SAMPLER_PSEUDO_NN_H_
-#define MUI_SAMPLER_PSEUDO_NN_H_
+#ifndef MUI_CHRONO_SAMPLER_EXACT_H_
+#define MUI_CHRONO_SAMPLER_EXACT_H_
 
-#include "config.h"
-#include "sampler.h"
+#include "../util.h"
+#include "../config.h"
 
 namespace mui {
 
-template<typename O_TP, typename I_TP=O_TP, typename CONFIG=default_config>
-class sampler_pseudo_nearest_neighbor {
+template<typename CONFIG=default_config> class chrono_sampler_exact {
 public:
-	using OTYPE      = O_TP;
-	using ITYPE      = I_TP;
 	using REAL       = typename CONFIG::REAL;
 	using INT        = typename CONFIG::INT;
-	using point_type = typename CONFIG::point_type;
+	using time_type  = typename CONFIG::time_type;
+	
+	chrono_sampler_exact( time_type tol = time_type(0) ) {
+		tolerance = tol;
+	}
 
-	sampler_pseudo_nearest_neighbor( REAL h_ ) : h(h_) {}
-
-	template<template<typename,typename> class CONTAINER>
-	inline OTYPE filter( point_type focus, const CONTAINER<ITYPE,CONFIG> &data_points ) const {
-		REAL r2min = std::numeric_limits<REAL>::max();
-		OTYPE value = 0;
-		for(INT i = 0 ; i < data_points.size() ; i++) {
-			REAL dr2 = normsq( focus - data_points[i].first );
-			if ( dr2 < r2min ) {
-				r2min = dr2;
-				value = data_points[i].second ;
+	template<typename TYPE>
+	TYPE filter( time_type focus, const std::vector<std::pair<time_type, TYPE> > &points ) const {
+		for( auto i: points ) {
+			if ( std::abs(i.first - focus) <= tolerance ) {
+				return i.second;
 			}
 		}
-		return value;
+		return TYPE(0);
 	}
-	inline geometry::any_shape<CONFIG> support( point_type focus ) const {
-		return geometry::sphere<CONFIG>( focus, h );
+	time_type get_upper_bound( time_type focus ) const {
+		return focus + tolerance;
 	}
+	time_type get_lower_bound( time_type focus ) const {
+		return focus - tolerance;
+	}
+
 protected:
-	REAL h;
+	time_type tolerance;
 };
 
 }
 
-#endif /* MUI_SAMPLER_NN_H_ */
+#endif /* MUI_CHRONO_SAMPLER_EXACT_H_ */
