+++ chrono_samplers/chrono_sampler_gauss.h
@@ -38,55 +38,56 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 ** File Details **
 
-Filename: sampler_gauss.h
-Created: Feb 10, 2014
+Filename: chrono_sampler_exact.h
+Created: Feb 19, 2014
 Author: Y. H. Tang
-Description: Spatial sampler that provides a value at a point
-             using Gaussian interpolation.
+Description: Temporal sampler that applies Gaussian interpolation
+             and is symmetric for past and future
 */
 
-#ifndef MUI_SAMPLER_GAUSS_H_
-#define MUI_SAMPLER_GAUSS_H_
+#ifndef MUI_SAMPLER_TIME_GAUSS_H_
+#define MUI_SAMPLER_TIME_GAUSS_H_
 
-#include "util.h"
+#include "../util.h"
+#include "../config.h"
 
 namespace mui {
 
-template<typename O_TP, typename I_TP=O_TP, typename CONFIG=default_config>
-class sampler_gauss {
+template<typename CONFIG=default_config> class chrono_sampler_gauss {
 public:
-	using OTYPE      = O_TP;
-	using ITYPE      = I_TP;
 	using REAL       = typename CONFIG::REAL;
 	using INT        = typename CONFIG::INT;
-	using point_type = typename CONFIG::point_type;
+	using time_type  = typename CONFIG::time_type;
+	
+	chrono_sampler_gauss( time_type newcutoff, REAL newsigma ) {
+		sigma  = newsigma;
+		cutoff = newcutoff;
+	}
 
-	sampler_gauss( REAL r_, REAL h_ ) : r(r_), h(h_), nh(std::pow(2*PI*h,-0.5*CONFIG::D)) {}
-
-	template<template<typename,typename> class CONTAINER>
-	inline OTYPE filter( point_type focus, const CONTAINER<ITYPE,CONFIG> &data_points ) const {
-		REAL  wsum = 0;
-		OTYPE vsum = 0;
-		for(INT i = 0 ; i < data_points.size() ; i++) {
-			auto d = normsq( focus - data_points[i].first );
-			if ( d < r*r ) {
-				REAL w = nh * std::exp( (-0.5/h) * d );
-				vsum += data_points[i].second * w;
+	template<typename TYPE>
+	TYPE filter( time_type focus, const std::vector<std::pair<time_type, TYPE> > &points ) const {
+		REAL wsum = REAL(0);
+		TYPE vsum = TYPE(0);
+		for( auto i: points ) {
+			time_type dt = std::abs(i.first - focus);
+			if ( dt < cutoff ) {
+				REAL w = pow( 2*PI*sigma, -0.5 ) * exp( -0.5 * dt * dt / sigma );
+				vsum += i.second * w;
 				wsum += w;
 			}
 		}
-		if ( wsum ) return vsum / wsum;
-		else return 0.;
+		return ( wsum > std::numeric_limits<REAL>::epsilon() ) ? ( vsum / wsum ) : TYPE(0);
 	}
-
-	inline geometry::any_shape<CONFIG> support( point_type focus ) const {
-		return geometry::sphere<CONFIG>( focus, r );
+	time_type get_upper_bound( time_type focus ) const {
+		return focus + cutoff;
+	}
+	time_type get_lower_bound( time_type focus ) const {
+		return focus - cutoff;
 	}
 
 protected:
-	REAL r;
-	REAL h;
-	REAL nh;
+	time_type cutoff;
+	REAL sigma;
 };
 
 }
