+++ chrono_samplers/chrono_sampler_mean.h
@@ -38,60 +38,54 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 ** File Details **
 
-Filename: sampler_pseudo_n2_linear.h
-Created: Oct 10, 2014
+Filename: sampler_time_mean.h
+Created: Oct 12, 2014
 Author: Y. H. Tang
-Description: Spatial sampler that provides a value at a point
-             using a pseudo linear nearest neighbour interpolation.
+Description: Temporal sampler that averages in time with a
+             range from [ now - left, now + right ]
 */
 
-#ifndef MUI_SAMPLER_PSEUDO_N2_LINEAR_H_
-#define MUI_SAMPLER_PSEUDO_N2_LINEAR_H_
+#ifndef MUI_SAMPLER_TIME_MEAN_H_
+#define MUI_SAMPLER_TIME_MEAN_H_
 
-#include "config.h"
-#include "sampler.h"
+#include "../util.h"
+#include "../config.h"
 
 namespace mui {
 
-template<typename O_TP, typename I_TP=O_TP, typename CONFIG=default_config>
-class sampler_pseudo_nearest2_linear {
+template<typename CONFIG=default_config> class chrono_sampler_mean {
 public:
-	using OTYPE      = O_TP;
-	using ITYPE      = I_TP;
 	using REAL       = typename CONFIG::REAL;
 	using INT        = typename CONFIG::INT;
-	using point_type = typename CONFIG::point_type;
+	using time_type  = typename CONFIG::time_type;
+	
+	chrono_sampler_mean( time_type newleft = time_type(0), time_type newright = time_type(0) ) {
+		left   = newleft;
+		right  = newright;
+	}
 
-	sampler_pseudo_nearest2_linear( REAL h_ ) : h(h_) {}
-
-	template<template<typename,typename> class CONTAINER>
-	inline OTYPE filter( point_type focus, const CONTAINER<ITYPE,CONFIG> &data_points ) const {
-		REAL r2min_1st = std::numeric_limits<REAL>::max();
-		REAL r2min_2nd = std::numeric_limits<REAL>::max();
-		OTYPE value_1st = 0, value_2nd = 0;
-		for(INT i = 0 ; i < data_points.size() ; i++) {
-			REAL dr2 = normsq( focus - data_points[i].first );
-			if ( dr2 < r2min_1st ) {
-				r2min_2nd = r2min_1st;
-				value_2nd = value_1st;
-				r2min_1st = dr2;
-				value_1st = data_points[i].second ;
-			} else if ( dr2 < r2min_2nd ) {
-				r2min_2nd = dr2;
-				value_2nd = data_points[i].second ;
+	template<typename TYPE>
+	TYPE filter( time_type focus, const std::vector<std::pair<time_type, TYPE> > &points ) const {
+		TYPE sum = TYPE(0);
+		for( auto i: points ) {
+			if ( i.first <= focus + right && i.first >= focus - left ) {
+				sum += i.second;
 			}
 		}
-		REAL r1 = std::sqrt( r2min_1st );
-		REAL r2 = std::sqrt( r2min_2nd );
-		return ( value_1st * r2 + value_2nd * r1 ) / ( r1 + r2 );
+		if ( points.size() ) return sum / TYPE(points.size());
+		else return TYPE(0);
 	}
-	inline geometry::any_shape<CONFIG> support( point_type focus ) const {
-		return geometry::sphere<CONFIG>( focus, h );
+	time_type get_upper_bound( time_type focus ) const {
+		return focus + right;
 	}
+	time_type get_lower_bound( time_type focus ) const {
+		return focus - left;
+	}
+
 protected:
-	REAL h;
+	time_type left, right;
 };
 
 }
 
-#endif /* MUI_SAMPLER_NN_H_ */
+#endif /* MUI_SAMPLER_TIME_SUM_H_ */
