+++ chrono_samplers/chrono_sampler_null.h
@@ -38,50 +38,43 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 ** File Details **
 
-Filename: sampler_nn.h
-Created: Feb 10, 2014
+Filename: chrono_sampler_null.h
+Created: Nov 5, 2014
 Author: Y. H. Tang
-Description: Spatial sampler that provides a value at a point
-             using a nearest neighbour interpolation.
+Description: Dummy temporal sampler intended as a file template for
+             creating new samplers
 */
 
-#ifndef MUI_SAMPLER_NN_H_
-#define MUI_SAMPLER_NN_H_
+#ifndef MUI_SAMPLER_TIME_NULL_H_
+#define MUI_SAMPLER_TIME_NULL_H_
 
-#include "config.h"
-#include "sampler.h"
+#include "../util.h"
+#include "../config.h"
 
 namespace mui {
 
-template<typename O_TP, typename I_TP=O_TP, typename CONFIG=default_config>
-class sampler_nearest_neighbor {
+template<typename CONFIG=default_config> class chrono_sampler_null {
 public:
-	using OTYPE      = O_TP;
-	using ITYPE      = I_TP;
 	using REAL       = typename CONFIG::REAL;
 	using INT        = typename CONFIG::INT;
-	using point_type = typename CONFIG::point_type;
-
-	sampler_nearest_neighbor() {}
-
-	template<template<typename,typename> class CONTAINER>
-	inline OTYPE filter( point_type focus, const CONTAINER<ITYPE,CONFIG> &data_points ) const {
-		REAL r2min = std::numeric_limits<REAL>::max();
-		OTYPE value = 0;
-		for(INT i = 0 ; i < data_points.size() ; i++) {
-			REAL dr2 = normsq( focus - data_points[i].first );
-			if ( dr2 < r2min ) {
-				r2min = dr2;
-				value = data_points[i].second ;
-			}
-		}
-		return value;
+	using time_type  = typename CONFIG::time_type;
+	
+	chrono_sampler_null() {
+		// to do: initialization
 	}
-	inline geometry::any_shape<CONFIG> support( point_type focus ) const {
-		return geometry::point<CONFIG>( focus );
+
+	template<typename TYPE>
+	TYPE filter( time_type focus, const std::vector<std::pair<time_type, TYPE> > &points ) const {
+		// to do: interpolation algorithm
+	}
+	time_type get_upper_bound( time_type focus ) const {
+		// to do: return newest time needed with regard to focus
+	}
+	time_type get_lower_bound( time_type focus ) const {
+		// to do: return oldest time needed with regard to focus
 	}
 };
 
 }
 
-#endif /* MUI_SAMPLER_NN_H_ */
+#endif /* MUI_SAMPLER_TIME_NULL_H_ */
