+++ spatial_samplers/sampler_pseudo_nn.h
@@ -38,58 +38,52 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 ** File Details **
 
-Filename: chrono_sampler_exact.h
-Created: Feb 19, 2014
+Filename: sampler_pseudo_nn.h
+Created: Oct 10, 2014
 Author: Y. H. Tang
-Description: Temporal sampler that applies Gaussian interpolation
-             and is symmetric for past and future
+Description: Spatial sampler that provides a value at a point
+             using a pseudo nearest neighbour interpolation.
 */
 
-#ifndef MUI_SAMPLER_TIME_GAUSS_H_
-#define MUI_SAMPLER_TIME_GAUSS_H_
+#ifndef MUI_SAMPLER_PSEUDO_NN_H_
+#define MUI_SAMPLER_PSEUDO_NN_H_
 
-#include "util.h"
-#include "config.h"
+#include "../config.h"
+#include "../sampler.h"
 
 namespace mui {
 
-template<typename CONFIG=default_config> class chrono_sampler_gauss {
+template<typename O_TP, typename I_TP=O_TP, typename CONFIG=default_config>
+class sampler_pseudo_nearest_neighbor {
 public:
+	using OTYPE      = O_TP;
+	using ITYPE      = I_TP;
 	using REAL       = typename CONFIG::REAL;
 	using INT        = typename CONFIG::INT;
-	using time_type  = typename CONFIG::time_type;
-	
-	chrono_sampler_gauss( time_type newcutoff, REAL newsigma ) {
-		sigma  = newsigma;
-		cutoff = newcutoff;
-	}
+	using point_type = typename CONFIG::point_type;
 
-	template<typename TYPE>
-	TYPE filter( time_type focus, const std::vector<std::pair<time_type, TYPE> > &points ) const {
-		REAL wsum = REAL(0);
-		TYPE vsum = TYPE(0);
-		for( auto i: points ) {
-			time_type dt = std::abs(i.first - focus);
-			if ( dt < cutoff ) {
-				REAL w = pow( 2*PI*sigma, -0.5 ) * exp( -0.5 * dt * dt / sigma );
-				vsum += i.second * w;
-				wsum += w;
+	sampler_pseudo_nearest_neighbor( REAL h_ ) : h(h_) {}
+
+	template<template<typename,typename> class CONTAINER>
+	inline OTYPE filter( point_type focus, const CONTAINER<ITYPE,CONFIG> &data_points ) const {
+		REAL r2min = std::numeric_limits<REAL>::max();
+		OTYPE value = 0;
+		for(INT i = 0 ; i < data_points.size() ; i++) {
+			REAL dr2 = normsq( focus - data_points[i].first );
+			if ( dr2 < r2min ) {
+				r2min = dr2;
+				value = data_points[i].second ;
 			}
 		}
-		return ( wsum > std::numeric_limits<REAL>::epsilon() ) ? ( vsum / wsum ) : TYPE(0);
+		return value;
 	}
-	time_type get_upper_bound( time_type focus ) const {
-		return focus + cutoff;
+	inline geometry::any_shape<CONFIG> support( point_type focus ) const {
+		return geometry::sphere<CONFIG>( focus, h );
 	}
-	time_type get_lower_bound( time_type focus ) const {
-		return focus - cutoff;
-	}
-
 protected:
-	time_type cutoff;
-	REAL sigma;
+	REAL h;
 };
 
 }
 
-#endif /* MUI_SAMPLER_GAUSS_H_ */
+#endif /* MUI_SAMPLER_NN_H_ */
